+++ src/Console/UpdateCommand.php
@@ -1,0 +1,42 @@
+<?php
+/**
+ * All rights reserved. No part of this code may be reproduced, modified,
+ * amended or retransmitted in any form or by any means for any purpose without
+ * prior written consent of Mizmoz Limited.
+ * You must ensure that this copyright notice remains intact at all times
+ *
+ * @copyright Copyright (c) Mizmoz Limited 2017. All rights reserved.
+ */
+
+namespace Mizmoz\Validate\Console;
+
+use Mizmoz\Validate\Console\Update\IsEmailDisposable;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class UpdateCommand extends Command
+{
+    /**
+     * @inheritdoc
+     */
+    protected function configure()
+    {
+        $this->setName('update')
+            ->setDescription('Update the data files for the validation');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $output->writeln('Updating the data files.');
+
+        // write the files to
+        $filePath = __DIR__ . '/../../resources/is-email-disposable.php';
+
+        // update the disposable lists
+        (new IsEmailDisposable())->update($filePath);
+    }
+}